@@ -6,7 +6,7 @@ description: Demonstrates how to use the diac_client plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -19,16 +19,16 @@ dependencies:
     # The example app is bundled with the plugin so we use a path dependency on
     # the parent directory to use the current plugin's version. 
     path: ../
-  logging: ^0.11.4
-  logging_appenders: ^0.4.1
-  uuid: ^2.0.4
-  flutter_async_utils: ^0.0.1
-  clock: ^1.0.1
-  url_launcher_web: ">=0.1.1+5 <2.0.0"
+  logging: ^1.0.1
+  logging_appenders: ^1.0.0
+  uuid: ^3.0.4
+  flutter_async_utils: ^1.0.0-nullsafety.1
+  clock: ^1.1.0
+  url_launcher_web: ^2.0.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.3
 
 dev_dependencies:
   flutter_test:
