@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
 
 part of 'diac_dto.dart';
 
@@ -8,10 +8,15 @@ part of 'diac_dto.dart';
 // **************************************************************************
 
 T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
 DiacConfig _$DiacConfigFromJson(Map<String, dynamic> json) {
   return _DiacConfig.fromJson(json);
 }
 
+/// @nodoc
 class _$DiacConfigTearOff {
   const _$DiacConfigTearOff();
 
@@ -22,30 +27,39 @@ class _$DiacConfigTearOff {
       messages: messages,
     );
   }
+
+  DiacConfig fromJson(Map<String, Object> json) {
+    return DiacConfig.fromJson(json);
+  }
 }
 
-// ignore: unused_element
+/// @nodoc
 const $DiacConfig = _$DiacConfigTearOff();
 
+/// @nodoc
 mixin _$DiacConfig {
-  DateTime get updatedAt;
-  List<DiacMessage> get messages;
+  DateTime get updatedAt => throw _privateConstructorUsedError;
+  List<DiacMessage> get messages => throw _privateConstructorUsedError;
 
-  Map<String, dynamic> toJson();
-  $DiacConfigCopyWith<DiacConfig> get copyWith;
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $DiacConfigCopyWith<DiacConfig> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
+/// @nodoc
 abstract class $DiacConfigCopyWith<$Res> {
   factory $DiacConfigCopyWith(
-          DiacConfig? value, $Res Function(DiacConfig) then) =
+          DiacConfig value, $Res Function(DiacConfig) then) =
       _$DiacConfigCopyWithImpl<$Res>;
-  $Res call({DateTime? updatedAt, List<DiacMessage>? messages});
+  $Res call({DateTime updatedAt, List<DiacMessage> messages});
 }
 
+/// @nodoc
 class _$DiacConfigCopyWithImpl<$Res> implements $DiacConfigCopyWith<$Res> {
   _$DiacConfigCopyWithImpl(this._value, this._then);
 
-  final DiacConfig? _value;
+  final DiacConfig _value;
   // ignore: unused_field
   final $Res Function(DiacConfig) _then;
 
@@ -54,24 +68,29 @@ class _$DiacConfigCopyWithImpl<$Res> implements $DiacConfigCopyWith<$Res> {
     Object? updatedAt = freezed,
     Object? messages = freezed,
   }) {
-    return _then(_value!.copyWith(
-      updatedAt:
-          updatedAt == freezed ? _value!.updatedAt : updatedAt as DateTime?,
+    return _then(_value.copyWith(
+      updatedAt: updatedAt == freezed
+          ? _value.updatedAt
+          : updatedAt // ignore: cast_nullable_to_non_nullable
+              as DateTime,
       messages: messages == freezed
-          ? _value!.messages
-          : messages as List<DiacMessage>?,
+          ? _value.messages
+          : messages // ignore: cast_nullable_to_non_nullable
+              as List<DiacMessage>,
     ));
   }
 }
 
+/// @nodoc
 abstract class _$DiacConfigCopyWith<$Res> implements $DiacConfigCopyWith<$Res> {
   factory _$DiacConfigCopyWith(
           _DiacConfig value, $Res Function(_DiacConfig) then) =
       __$DiacConfigCopyWithImpl<$Res>;
   @override
-  $Res call({DateTime? updatedAt, List<DiacMessage>? messages});
+  $Res call({DateTime updatedAt, List<DiacMessage> messages});
 }
 
+/// @nodoc
 class __$DiacConfigCopyWithImpl<$Res> extends _$DiacConfigCopyWithImpl<$Res>
     implements _$DiacConfigCopyWith<$Res> {
   __$DiacConfigCopyWithImpl(
@@ -79,7 +98,7 @@ class __$DiacConfigCopyWithImpl<$Res> extends _$DiacConfigCopyWithImpl<$Res>
       : super(_value, (v) => _then(v as _DiacConfig));
 
   @override
-  _DiacConfig? get _value => super._value as _DiacConfig?;
+  _DiacConfig get _value => super._value as _DiacConfig;
 
   @override
   $Res call({
@@ -87,20 +106,22 @@ class __$DiacConfigCopyWithImpl<$Res> extends _$DiacConfigCopyWithImpl<$Res>
     Object? messages = freezed,
   }) {
     return _then(_DiacConfig(
-      updatedAt:
-          updatedAt == freezed ? _value!.updatedAt : (updatedAt as DateTime?)!,
+      updatedAt: updatedAt == freezed
+          ? _value.updatedAt
+          : updatedAt // ignore: cast_nullable_to_non_nullable
+              as DateTime,
       messages: messages == freezed
-          ? _value!.messages
-          : (messages as List<DiacMessage>?)!,
+          ? _value.messages
+          : messages // ignore: cast_nullable_to_non_nullable
+              as List<DiacMessage>,
     ));
   }
 }
 
+/// @nodoc
 @JsonSerializable()
 class _$_DiacConfig implements _DiacConfig {
-  const _$_DiacConfig({required this.updatedAt, required this.messages})
-      : assert(updatedAt != null),
-        assert(messages != null);
+  const _$_DiacConfig({required this.updatedAt, required this.messages});
 
   factory _$_DiacConfig.fromJson(Map<String, dynamic> json) =>
       _$_$_DiacConfigFromJson(json);
@@ -133,6 +154,7 @@ class _$_DiacConfig implements _DiacConfig {
       const DeepCollectionEquality().hash(updatedAt) ^
       const DeepCollectionEquality().hash(messages);
 
+  @JsonKey(ignore: true)
   @override
   _$DiacConfigCopyWith<_DiacConfig> get copyWith =>
       __$DiacConfigCopyWithImpl<_DiacConfig>(this, _$identity);
@@ -152,17 +174,20 @@ abstract class _DiacConfig implements DiacConfig {
       _$_DiacConfig.fromJson;
 
   @override
-  DateTime get updatedAt;
+  DateTime get updatedAt => throw _privateConstructorUsedError;
   @override
-  List<DiacMessage> get messages;
+  List<DiacMessage> get messages => throw _privateConstructorUsedError;
   @override
-  _$DiacConfigCopyWith<_DiacConfig> get copyWith;
+  @JsonKey(ignore: true)
+  _$DiacConfigCopyWith<_DiacConfig> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 DiacMessage _$DiacMessageFromJson(Map<String, dynamic> json) {
   return _DiacMessage.fromJson(json);
 }
 
+/// @nodoc
 class _$DiacMessageTearOff {
   const _$DiacMessageTearOff();
 
@@ -184,38 +209,47 @@ class _$DiacMessageTearOff {
       actions: actions,
     );
   }
+
+  DiacMessage fromJson(Map<String, Object> json) {
+    return DiacMessage.fromJson(json);
+  }
 }
 
-// ignore: unused_element
+/// @nodoc
 const $DiacMessage = _$DiacMessageTearOff();
 
+/// @nodoc
 mixin _$DiacMessage {
-  String get uuid;
-  String get body;
-  String get key;
-  DateTime? get dateStart;
-  DateTime? get dateEnd;
-  String? get expression;
-  List<DiacMessageAction> get actions;
+  String get uuid => throw _privateConstructorUsedError;
+  String get body => throw _privateConstructorUsedError;
+  String get key => throw _privateConstructorUsedError;
+  DateTime? get dateStart => throw _privateConstructorUsedError;
+  DateTime? get dateEnd => throw _privateConstructorUsedError;
+  String? get expression => throw _privateConstructorUsedError;
+  List<DiacMessageAction> get actions => throw _privateConstructorUsedError;
 
-  Map<String, dynamic> toJson();
-  $DiacMessageCopyWith<DiacMessage> get copyWith;
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $DiacMessageCopyWith<DiacMessage> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
+/// @nodoc
 abstract class $DiacMessageCopyWith<$Res> {
   factory $DiacMessageCopyWith(
           DiacMessage value, $Res Function(DiacMessage) then) =
       _$DiacMessageCopyWithImpl<$Res>;
   $Res call(
-      {String? uuid,
-      String? body,
-      String? key,
+      {String uuid,
+      String body,
+      String key,
       DateTime? dateStart,
       DateTime? dateEnd,
       String? expression,
-      List<DiacMessageAction>? actions});
+      List<DiacMessageAction> actions});
 }
 
+/// @nodoc
 class _$DiacMessageCopyWithImpl<$Res> implements $DiacMessageCopyWith<$Res> {
   _$DiacMessageCopyWithImpl(this._value, this._then);
 
@@ -234,21 +268,39 @@ class _$DiacMessageCopyWithImpl<$Res> implements $DiacMessageCopyWith<$Res> {
     Object? actions = freezed,
   }) {
     return _then(_value.copyWith(
-      uuid: uuid == freezed ? _value.uuid : uuid as String?,
-      body: body == freezed ? _value.body : body as String?,
-      key: key == freezed ? _value.key : key as String?,
-      dateStart:
-          dateStart == freezed ? _value.dateStart : dateStart as DateTime?,
-      dateEnd: dateEnd == freezed ? _value.dateEnd : dateEnd as DateTime?,
-      expression:
-          expression == freezed ? _value.expression : expression as String?,
+      uuid: uuid == freezed
+          ? _value.uuid
+          : uuid // ignore: cast_nullable_to_non_nullable
+              as String,
+      body: body == freezed
+          ? _value.body
+          : body // ignore: cast_nullable_to_non_nullable
+              as String,
+      key: key == freezed
+          ? _value.key
+          : key // ignore: cast_nullable_to_non_nullable
+              as String,
+      dateStart: dateStart == freezed
+          ? _value.dateStart
+          : dateStart // ignore: cast_nullable_to_non_nullable
+              as DateTime?,
+      dateEnd: dateEnd == freezed
+          ? _value.dateEnd
+          : dateEnd // ignore: cast_nullable_to_non_nullable
+              as DateTime?,
+      expression: expression == freezed
+          ? _value.expression
+          : expression // ignore: cast_nullable_to_non_nullable
+              as String?,
       actions: actions == freezed
           ? _value.actions
-          : actions as List<DiacMessageAction>?,
+          : actions // ignore: cast_nullable_to_non_nullable
+              as List<DiacMessageAction>,
     ));
   }
 }
 
+/// @nodoc
 abstract class _$DiacMessageCopyWith<$Res>
     implements $DiacMessageCopyWith<$Res> {
   factory _$DiacMessageCopyWith(
@@ -256,15 +308,16 @@ abstract class _$DiacMessageCopyWith<$Res>
       __$DiacMessageCopyWithImpl<$Res>;
   @override
   $Res call(
-      {String? uuid,
-      String? body,
-      String? key,
+      {String uuid,
+      String body,
+      String key,
       DateTime? dateStart,
       DateTime? dateEnd,
       String? expression,
-      List<DiacMessageAction>? actions});
+      List<DiacMessageAction> actions});
 }
 
+/// @nodoc
 class __$DiacMessageCopyWithImpl<$Res> extends _$DiacMessageCopyWithImpl<$Res>
     implements _$DiacMessageCopyWith<$Res> {
   __$DiacMessageCopyWithImpl(
@@ -285,21 +338,39 @@ class __$DiacMessageCopyWithImpl<$Res> extends _$DiacMessageCopyWithImpl<$Res>
     Object? actions = freezed,
   }) {
     return _then(_DiacMessage(
-      uuid: uuid == freezed ? _value.uuid : (uuid as String?)!,
-      body: body == freezed ? _value.body : (body as String?)!,
-      key: key == freezed ? _value.key : (key as String?)!,
-      dateStart:
-          dateStart == freezed ? _value.dateStart : dateStart as DateTime?,
-      dateEnd: dateEnd == freezed ? _value.dateEnd : dateEnd as DateTime?,
-      expression:
-          expression == freezed ? _value.expression : expression as String?,
+      uuid: uuid == freezed
+          ? _value.uuid
+          : uuid // ignore: cast_nullable_to_non_nullable
+              as String,
+      body: body == freezed
+          ? _value.body
+          : body // ignore: cast_nullable_to_non_nullable
+              as String,
+      key: key == freezed
+          ? _value.key
+          : key // ignore: cast_nullable_to_non_nullable
+              as String,
+      dateStart: dateStart == freezed
+          ? _value.dateStart
+          : dateStart // ignore: cast_nullable_to_non_nullable
+              as DateTime?,
+      dateEnd: dateEnd == freezed
+          ? _value.dateEnd
+          : dateEnd // ignore: cast_nullable_to_non_nullable
+              as DateTime?,
+      expression: expression == freezed
+          ? _value.expression
+          : expression // ignore: cast_nullable_to_non_nullable
+              as String?,
       actions: actions == freezed
           ? _value.actions
-          : (actions as List<DiacMessageAction>?)!,
+          : actions // ignore: cast_nullable_to_non_nullable
+              as List<DiacMessageAction>,
     ));
   }
 }
 
+/// @nodoc
 @JsonSerializable()
 class _$_DiacMessage implements _DiacMessage {
   const _$_DiacMessage(
@@ -309,11 +380,7 @@ class _$_DiacMessage implements _DiacMessage {
       this.dateStart,
       this.dateEnd,
       this.expression,
-      required this.actions})
-      : assert(uuid != null),
-        assert(body != null),
-        assert(key != null),
-        assert(actions != null);
+      required this.actions});
 
   factory _$_DiacMessage.fromJson(Map<String, dynamic> json) =>
       _$_$_DiacMessageFromJson(json);
@@ -372,6 +439,7 @@ class _$_DiacMessage implements _DiacMessage {
       const DeepCollectionEquality().hash(expression) ^
       const DeepCollectionEquality().hash(actions);
 
+  @JsonKey(ignore: true)
   @override
   _$DiacMessageCopyWith<_DiacMessage> get copyWith =>
       __$DiacMessageCopyWithImpl<_DiacMessage>(this, _$identity);
@@ -396,27 +464,30 @@ abstract class _DiacMessage implements DiacMessage {
       _$_DiacMessage.fromJson;
 
   @override
-  String get uuid;
+  String get uuid => throw _privateConstructorUsedError;
   @override
-  String get body;
+  String get body => throw _privateConstructorUsedError;
   @override
-  String get key;
+  String get key => throw _privateConstructorUsedError;
   @override
-  DateTime? get dateStart;
+  DateTime? get dateStart => throw _privateConstructorUsedError;
   @override
-  DateTime? get dateEnd;
+  DateTime? get dateEnd => throw _privateConstructorUsedError;
   @override
-  String? get expression;
+  String? get expression => throw _privateConstructorUsedError;
   @override
-  List<DiacMessageAction> get actions;
+  List<DiacMessageAction> get actions => throw _privateConstructorUsedError;
   @override
-  _$DiacMessageCopyWith<_DiacMessage> get copyWith;
+  @JsonKey(ignore: true)
+  _$DiacMessageCopyWith<_DiacMessage> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 DiacMessageAction _$DiacMessageActionFromJson(Map<String, dynamic> json) {
   return _DiacMessageAction.fromJson(json);
 }
 
+/// @nodoc
 class _$DiacMessageActionTearOff {
   const _$DiacMessageActionTearOff();
 
@@ -432,28 +503,39 @@ class _$DiacMessageActionTearOff {
       expression: expression,
     );
   }
+
+  DiacMessageAction fromJson(Map<String, Object> json) {
+    return DiacMessageAction.fromJson(json);
+  }
 }
 
-// ignore: unused_element
+/// @nodoc
 const $DiacMessageAction = _$DiacMessageActionTearOff();
 
+/// @nodoc
 mixin _$DiacMessageAction {
-  String get key;
-  String get label;
-  String? get url;
-  String? get expression;
+  String get key => throw _privateConstructorUsedError;
+  String get label => throw _privateConstructorUsedError;
+  String? get url => throw _privateConstructorUsedError;
 
-  Map<String, dynamic> toJson();
-  $DiacMessageActionCopyWith<DiacMessageAction> get copyWith;
+  /// Expression which will be evaluated, if url is null.
+  String? get expression => throw _privateConstructorUsedError;
+
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $DiacMessageActionCopyWith<DiacMessageAction> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
+/// @nodoc
 abstract class $DiacMessageActionCopyWith<$Res> {
   factory $DiacMessageActionCopyWith(
           DiacMessageAction value, $Res Function(DiacMessageAction) then) =
       _$DiacMessageActionCopyWithImpl<$Res>;
-  $Res call({String? key, String? label, String? url, String? expression});
+  $Res call({String key, String label, String? url, String? expression});
 }
 
+/// @nodoc
 class _$DiacMessageActionCopyWithImpl<$Res>
     implements $DiacMessageActionCopyWith<$Res> {
   _$DiacMessageActionCopyWithImpl(this._value, this._then);
@@ -470,24 +552,37 @@ class _$DiacMessageActionCopyWithImpl<$Res>
     Object? expression = freezed,
   }) {
     return _then(_value.copyWith(
-      key: key == freezed ? _value.key : key as String?,
-      label: label == freezed ? _value.label : label as String?,
-      url: url == freezed ? _value.url : url as String?,
-      expression:
-          expression == freezed ? _value.expression : expression as String?,
+      key: key == freezed
+          ? _value.key
+          : key // ignore: cast_nullable_to_non_nullable
+              as String,
+      label: label == freezed
+          ? _value.label
+          : label // ignore: cast_nullable_to_non_nullable
+              as String,
+      url: url == freezed
+          ? _value.url
+          : url // ignore: cast_nullable_to_non_nullable
+              as String?,
+      expression: expression == freezed
+          ? _value.expression
+          : expression // ignore: cast_nullable_to_non_nullable
+              as String?,
     ));
   }
 }
 
+/// @nodoc
 abstract class _$DiacMessageActionCopyWith<$Res>
     implements $DiacMessageActionCopyWith<$Res> {
   factory _$DiacMessageActionCopyWith(
           _DiacMessageAction value, $Res Function(_DiacMessageAction) then) =
       __$DiacMessageActionCopyWithImpl<$Res>;
   @override
-  $Res call({String? key, String? label, String? url, String? expression});
+  $Res call({String key, String label, String? url, String? expression});
 }
 
+/// @nodoc
 class __$DiacMessageActionCopyWithImpl<$Res>
     extends _$DiacMessageActionCopyWithImpl<$Res>
     implements _$DiacMessageActionCopyWith<$Res> {
@@ -506,21 +601,31 @@ class __$DiacMessageActionCopyWithImpl<$Res>
     Object? expression = freezed,
   }) {
     return _then(_DiacMessageAction(
-      key: key == freezed ? _value.key : (key as String?)!,
-      label: label == freezed ? _value.label : (label as String?)!,
-      url: url == freezed ? _value.url : url as String?,
-      expression:
-          expression == freezed ? _value.expression : expression as String?,
+      key: key == freezed
+          ? _value.key
+          : key // ignore: cast_nullable_to_non_nullable
+              as String,
+      label: label == freezed
+          ? _value.label
+          : label // ignore: cast_nullable_to_non_nullable
+              as String,
+      url: url == freezed
+          ? _value.url
+          : url // ignore: cast_nullable_to_non_nullable
+              as String?,
+      expression: expression == freezed
+          ? _value.expression
+          : expression // ignore: cast_nullable_to_non_nullable
+              as String?,
     ));
   }
 }
 
+/// @nodoc
 @JsonSerializable()
 class _$_DiacMessageAction implements _DiacMessageAction {
   const _$_DiacMessageAction(
-      {required this.key, required this.label, this.url, this.expression})
-      : assert(key != null),
-        assert(label != null);
+      {required this.key, required this.label, this.url, this.expression});
 
   factory _$_DiacMessageAction.fromJson(Map<String, dynamic> json) =>
       _$_$_DiacMessageActionFromJson(json);
@@ -532,6 +637,8 @@ class _$_DiacMessageAction implements _DiacMessageAction {
   @override
   final String? url;
   @override
+
+  /// Expression which will be evaluated, if url is null.
   final String? expression;
 
   @override
@@ -562,6 +669,7 @@ class _$_DiacMessageAction implements _DiacMessageAction {
       const DeepCollectionEquality().hash(url) ^
       const DeepCollectionEquality().hash(expression);
 
+  @JsonKey(ignore: true)
   @override
   _$DiacMessageActionCopyWith<_DiacMessageAction> get copyWith =>
       __$DiacMessageActionCopyWithImpl<_DiacMessageAction>(this, _$identity);
@@ -583,13 +691,17 @@ abstract class _DiacMessageAction implements DiacMessageAction {
       _$_DiacMessageAction.fromJson;
 
   @override
-  String get key;
+  String get key => throw _privateConstructorUsedError;
   @override
-  String get label;
+  String get label => throw _privateConstructorUsedError;
   @override
-  String? get url;
+  String? get url => throw _privateConstructorUsedError;
   @override
-  String? get expression;
+
+  /// Expression which will be evaluated, if url is null.
+  String? get expression => throw _privateConstructorUsedError;
   @override
-  _$DiacMessageActionCopyWith<_DiacMessageAction> get copyWith;
+  @JsonKey(ignore: true)
+  _$DiacMessageActionCopyWith<_DiacMessageAction> get copyWith =>
+      throw _privateConstructorUsedError;
 }
